LayoutTests/fast/selectors/attribute-hyphen-value-matching-is-ascii-case-insensitive-expected.txt

When matching attributes case insensitively, it should be ASCII case insensitive. This test verifies the behavior when doing hyphen matching on the value (e.g. [a|="b"])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "[data-attribute|=Web-É-Kit]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É-Kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É-Kit]')[0].id is "target1"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=Web-É]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É]')[0].id is "target1"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=web-É]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=web-É]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=WEB-É-KIT]"
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É-KIT]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É-KIT]')[0].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=WEB-É]"
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É]')[0].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=Web-é]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-é]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=Web-é]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=web-é-kit]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-é-kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=web-é-kit]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute|=web-é]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-é]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=web-é]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute|=Web-É-Kit i]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É-Kit i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É-Kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É-Kit i]')[1].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=Web-É i]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=web-É i]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute|=web-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute|=web-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute|=web-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=WEB-É i]"
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute|=WEB-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute|=Web-é i]"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-é i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute|=Web-é i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute|=Web-é i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute|=web-é-kit i]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-é-kit i]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute|=web-é-kit i]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute|=web-é i]"
PASS document.querySelectorAll('#test-root [data-attribute|=web-é i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute|=web-é i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute|=web-é i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple|=WEB-é-KIT]"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é-KIT]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=WEB-é-KIT]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=WEB-é]"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=WEB-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é-Kit]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é-Kit]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=Web-é-Kit]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é-kit]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é-kit]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=Web-é-kit]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=Web-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É-kit]"
PASS document.querySelectorAll('#test-root [multiple|=web-É-kit]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=web-É-kit]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É]"
PASS document.querySelectorAll('#test-root [multiple|=web-É]').length is 2
PASS document.querySelectorAll('#test-root [multiple|=web-É]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple|=web-É]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple|=web-É-Kit]"
PASS document.querySelectorAll('#test-root [multiple|=web-É-Kit]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=web-É-Kit]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É]"
PASS document.querySelectorAll('#test-root [multiple|=web-É]').length is 2
PASS document.querySelectorAll('#test-root [multiple|=web-É]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple|=web-É]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple|=WEB-é-KIT i]"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é-KIT i]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=WEB-é-KIT i]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=WEB-é i]"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=WEB-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=WEB-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é-Kit i]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é-Kit i]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=Web-é-Kit i]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é i]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é-kit i]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é-kit i]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=Web-é-kit i]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=Web-é i]"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple|=Web-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É-kit i]"
PASS document.querySelectorAll('#test-root [multiple|=web-É-kit i]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=web-É-kit i]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É i]"
PASS document.querySelectorAll('#test-root [multiple|=web-É i]').length is 2
PASS document.querySelectorAll('#test-root [multiple|=web-É i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple|=web-É i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple|=web-É-Kit i]"
PASS document.querySelectorAll('#test-root [multiple|=web-É-Kit i]').length is 1
PASS document.querySelectorAll('#test-root [multiple|=web-É-Kit i]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple|=web-É i]"
PASS document.querySelectorAll('#test-root [multiple|=web-É i]').length is 2
PASS document.querySelectorAll('#test-root [multiple|=web-É i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple|=web-É i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

PASS successfullyParsed is true

TEST COMPLETE
